Backup of Nora EFI - 12.30.2023
Update to 0.9.8 and latest commits of each kext.
Delete AstekFusion2Adapter and AstekFusion2Family as the hardware is not in Nora anymore.
Delete AirportBrcmFixup as this is not needed on Apple AirPort wireless cards.
Add CatalinaI120Ethernet to prevent panics with attaching Intel I-120 Ethernet chipsets.
Add self-patched SSDT tables for various system functions.
Delete SSDT-GPU-DISABLE, as it is not in use.
Fix USBMap on iMacPro1,1 (I forgot it was set to MacPro7,1.)
Update DeviceProperties to match swapped GPUs and wireless card slots.
Update README to fix the rest of its information.